PR fixes
- Moved dataclasses to another file
- Moved "http://", "https://" strings to variable and used it throughout the code
- Remove `_get_mapping_meta` method. Instead, get it directly from _MappingMeta dataclass
