+++ spinta/__init__.py
@@ -1,3 +1,5 @@
 import importlib.metadata
 
 __version__ = importlib.metadata.version(__name__)
+
+HTTP_URL_PREFIXES = ("http://", "https://")
